+++ oshi-core/src/main/java/oshi/SystemInfo.java
@@ -23,6 +23,15 @@
  */
 package oshi;
 
+import static oshi.PlatformEnum.AIX;
+import static oshi.PlatformEnum.FREEBSD;
+import static oshi.PlatformEnum.LINUX;
+import static oshi.PlatformEnum.MACOS;
+import static oshi.PlatformEnum.MACOSX;
+import static oshi.PlatformEnum.OPENBSD;
+import static oshi.PlatformEnum.SOLARIS;
+import static oshi.PlatformEnum.UNKNOWN;
+import static oshi.PlatformEnum.WINDOWS;
 import static oshi.util.Memoizer.memoize;
 
 import java.util.function.Supplier;
@@ -47,7 +56,7 @@ import oshi.software.os.unix.solaris.SolarisOperatingSystem;
 import oshi.software.os.windows.WindowsOperatingSystem;
 
 /**
- * System information. This is the main entry point to Oshi.
+ * System information. This is the main entry point to OSHI.
  * <p>
  * This object provides getters which instantiate the appropriate
  * platform-specific implementations of {@link oshi.software.os.OperatingSystem}
@@ -57,25 +66,25 @@ public class SystemInfo {
 
     // The platform isn't going to change, and making this static enables easy
     // access from outside this class
-    private static final PlatformEnum currentPlatformEnum;
+    private static final PlatformEnum currentPlatform;
 
     static {
         if (Platform.isWindows()) {
-            currentPlatformEnum = PlatformEnum.WINDOWS;
+            currentPlatform = WINDOWS;
         } else if (Platform.isLinux()) {
-            currentPlatformEnum = PlatformEnum.LINUX;
+            currentPlatform = LINUX;
         } else if (Platform.isMac()) {
-            currentPlatformEnum = PlatformEnum.MACOSX;
+            currentPlatform = MACOS;
         } else if (Platform.isSolaris()) {
-            currentPlatformEnum = PlatformEnum.SOLARIS;
+            currentPlatform = SOLARIS;
         } else if (Platform.isFreeBSD()) {
-            currentPlatformEnum = PlatformEnum.FREEBSD;
+            currentPlatform = FREEBSD;
         } else if (Platform.isAIX()) {
-            currentPlatformEnum = PlatformEnum.AIX;
+            currentPlatform = AIX;
         } else if (Platform.isOpenBSD()) {
-            currentPlatformEnum = PlatformEnum.OPENBSD;
+            currentPlatform = OPENBSD;
         } else {
-            currentPlatformEnum = PlatformEnum.UNKNOWN;
+            currentPlatform = UNKNOWN;
         }
     }
 
@@ -84,14 +93,24 @@ public class SystemInfo {
     private final Supplier<HardwareAbstractionLayer> hardware = memoize(SystemInfo::createHardware);
 
     /**
-     * <p>
-     * Getter for the field <code>currentPlatformEnum</code>.
-     * </p>
+     * Gets the {@link PlatformEnum} value representing this system.
      *
-     * @return Returns the currentPlatformEnum.
+     * @return Returns the current platform
      */
+    public static PlatformEnum getCurrentPlatform() {
+        return currentPlatform;
+    }
+
+    /**
+     * Gets the {@link PlatformEnum} value representing this system.
+     *
+     * @return Returns the current platform
+     * @deprecated Use {@link #getCurrentPlatform()}
+     */
+    @Deprecated
     public static PlatformEnum getCurrentPlatformEnum() {
-        return currentPlatformEnum;
+        PlatformEnum platform = getCurrentPlatform();
+        return platform.equals(MACOS) ? MACOSX : platform;
     }
 
     /**
@@ -105,13 +124,12 @@ public class SystemInfo {
     }
 
     private static OperatingSystem createOperatingSystem() {
-        switch (currentPlatformEnum) {
-
+        switch (currentPlatform) {
         case WINDOWS:
             return new WindowsOperatingSystem();
         case LINUX:
             return new LinuxOperatingSystem();
-        case MACOSX:
+        case MACOS:
             return new MacOperatingSystem();
         case SOLARIS:
             return new SolarisOperatingSystem();
@@ -122,7 +140,8 @@ public class SystemInfo {
         case OPENBSD:
             return new OpenBsdOperatingSystem();
         default:
-            throw new UnsupportedOperationException("Operating system not supported: " + Platform.getOSType());
+            throw new UnsupportedOperationException(
+                    "Operating system not supported: JNA Platform type " + Platform.getOSType());
         }
     }
 
@@ -137,13 +156,12 @@ public class SystemInfo {
     }
 
     private static HardwareAbstractionLayer createHardware() {
-        switch (currentPlatformEnum) {
-
+        switch (currentPlatform) {
         case WINDOWS:
             return new WindowsHardwareAbstractionLayer();
         case LINUX:
             return new LinuxHardwareAbstractionLayer();
-        case MACOSX:
+        case MACOS:
             return new MacHardwareAbstractionLayer();
         case SOLARIS:
             return new SolarisHardwareAbstractionLayer();
@@ -154,7 +172,8 @@ public class SystemInfo {
         case OPENBSD:
             return new OpenBsdHardwareAbstractionLayer();
         default:
-            throw new UnsupportedOperationException("Operating system not supported: " + Platform.getOSType());
+            throw new UnsupportedOperationException(
+                    "Operating system not supported: JNA Platform type " + Platform.getOSType());
         }
     }
 }